+++ miscellaneoustool/gradle/libs.versions.toml
@@ -29,7 +29,7 @@ lifecycle-common = "2.6.1"
 lottie = "6.3.0"
 material = "1.11.0"
 mockk = "1.13.5"
-navigation = "2.8.0"
+navigation = "2.8.7"
 okhttp3 = "4.12.0"
 playInApp = "2.1.0"
 protobuf = "3.25.0"
